code improvements of header, and layouts
